Refactor RenderConfig out of RenderFrom{Endpoint,File}
