+++ demo-app/src/components/RenderFromEndpoint.tsx
@@ -1,13 +1,5 @@
 import React from 'react'
-import {
-  applyConfiguration,
-  parseYaml,
-  type COMPONENTS,
-} from '@open-truss/open-truss'
-
-// TODO: Set COMPONENT_INDEX in application config and OT loads it?
-import * as _components from '@/open-truss/components'
-const components = _components as unknown as COMPONENTS
+import RenderConfig from './RenderConfig'
 
 // TODO: Get this path from application config and only need to pass in filename?
 const CONFIG_API = '/ot/api/configs/'
@@ -20,7 +12,7 @@ export default function RenderFromEndpoint({
   // TODO: Use UQI's REST client once that exists?
   const url = `${CONFIG_API}${configName}`
   const [config, setConfig] = React.useState<string | null>(null)
-  const [loading, setLoading] = React.useState<boolean>(false)
+  const [loading, setLoading] = React.useState<boolean>(true)
   const [error, setError] = React.useState<Error | null>(null)
   React.useEffect(() => {
     const fetchConfig = async (): Promise<void> => {
@@ -35,23 +27,13 @@ export default function RenderFromEndpoint({
     setLoading(false)
   }, [url])
 
-  const configurationFunction = React.useMemo(
-    () => applyConfiguration(components),
-    [],
-  )
-
-  const renderedComponents = React.useMemo(() => {
-    if (config) {
-      const parsedConfig = parseYaml(config)
-      return configurationFunction(parsedConfig, {})
-    }
-  }, [config])
-
   if (error) {
     return <>OOPS! {error.message}</>
   } else if (loading) {
     return <>Loading...</>
+  } else if (config) {
+    return <RenderConfig config={config} />
   }
 
-  return <>{renderedComponents}</>
+  return <div>No config :(</div>
 }
